+++ src/test/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/RedshiftDataApiRepositoryTest.kt
@@ -75,6 +75,7 @@ class RedshiftDataApiRepositoryTest {
   private val executeStatementResponse = mock<ExecuteStatementResponse>()
   private val report = mock<Report>()
   private val datasetHelper = DatasetHelper()
+  private val redShiftSummaryTableHelper = mock<RedShiftSummaryTableHelper>()
 
   @BeforeEach
   fun setup() {
@@ -92,6 +93,7 @@ class RedshiftDataApiRepositoryTest {
       redshiftDataClient,
       tableIdGenerator,
       datasetHelper,
+      redShiftSummaryTableHelper,
       REDSHIFT_DATA_API_DB,
       REDSHIFT_DATA_API_CLUSTER_ID,
       REDSHIFT_DATA_API_SECRET_ARN,
@@ -147,6 +149,7 @@ SELECT *
       redshiftDataClient,
       tableIdGenerator,
       datasetHelper,
+      redShiftSummaryTableHelper,
       REDSHIFT_DATA_API_DB,
       REDSHIFT_DATA_API_CLUSTER_ID,
       REDSHIFT_DATA_API_SECRET_ARN,
@@ -188,6 +191,7 @@ SELECT *
       redshiftDataClient,
       tableIdGenerator,
       datasetHelper,
+      redShiftSummaryTableHelper,
       REDSHIFT_DATA_API_DB,
       REDSHIFT_DATA_API_CLUSTER_ID,
       REDSHIFT_DATA_API_SECRET_ARN,
@@ -231,6 +235,7 @@ SELECT *
       redshiftDataClient,
       tableIdGenerator,
       datasetHelper,
+      redShiftSummaryTableHelper,
       REDSHIFT_DATA_API_DB,
       REDSHIFT_DATA_API_CLUSTER_ID,
       REDSHIFT_DATA_API_SECRET_ARN,
@@ -285,6 +290,7 @@ SELECT *
       redshiftDataClient,
       tableIdGenerator,
       datasetHelper,
+      redShiftSummaryTableHelper,
       REDSHIFT_DATA_API_DB,
       REDSHIFT_DATA_API_CLUSTER_ID,
       REDSHIFT_DATA_API_SECRET_ARN,
@@ -373,6 +379,7 @@ SELECT *
       redshiftDataClient,
       tableIdGenerator,
       datasetHelper,
+      redShiftSummaryTableHelper,
       REDSHIFT_DATA_API_DB,
       REDSHIFT_DATA_API_CLUSTER_ID,
       REDSHIFT_DATA_API_SECRET_ARN,
@@ -400,6 +407,7 @@ SELECT *
       redshiftDataClient,
       tableIdGenerator,
       datasetHelper,
+      redShiftSummaryTableHelper,
       REDSHIFT_DATA_API_DB,
       REDSHIFT_DATA_API_CLUSTER_ID,
       REDSHIFT_DATA_API_SECRET_ARN,
@@ -425,6 +433,7 @@ SELECT *
       mock(),
       mock(),
       datasetHelper,
+      redShiftSummaryTableHelper,
       REDSHIFT_DATA_API_DB,
       REDSHIFT_DATA_API_CLUSTER_ID,
       REDSHIFT_DATA_API_SECRET_ARN,