+++ src/test/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/service/ConfiguredApiServiceTest.kt
@@ -1,7 +1,6 @@
 package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service
 
 import jakarta.validation.ValidationException
-import kotlinx.coroutines.test.runTest
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Assertions.assertEquals
@@ -1716,7 +1715,7 @@ class ConfiguredApiServiceTest {
   }
 
   @Test
-  fun `should call the repository with all provided arguments when getSummaryResult is called`() = runTest {
+  fun `should call the repository with all provided arguments when getSummaryResult is called`() {
     val tableId = TableIdGenerator().generateNewExternalTableId()
     val summaryId = "summaryId"
     whenever(
@@ -1734,7 +1733,7 @@ class ConfiguredApiServiceTest {
   }
 
   @Test
-  fun `should create and query summary table when it doesn't exist`() = runTest {
+  fun `should create and query summary table when it doesn't exist`() {
     val tableId = TableIdGenerator().generateNewExternalTableId()
     val summaryId = "summaryId"
     whenever(
@@ -1752,7 +1751,7 @@ class ConfiguredApiServiceTest {
 
     assertEquals(listOf(mapOf("total" to 1)), actual)
     verify(redshiftDataApiRepository, times(2)).getFullExternalTableResult(any(), anyOrNull())
-    verify(redshiftDataApiRepository).createSummaryTable(any(), any(), any(), any())
+    verify(configuredApiRepository).createSummaryTable(any(), any(), any(), any())
   }
 
   @Test